UI: Add react-number-format lib, rename Budget2 class
use format library to format amounts while typing,
delete deprecated Budget.js and replace it with Budget2,
fix imports
