+++ frontend/src/pages/Common/NumberFormat.js
@@ -1,0 +1,24 @@
+import React from "react";
+
+import NumberFormat from "react-number-format";
+
+export function applyNumberFormat(props) {
+  const { inputRef, onChange, ...other } = props;
+
+  return (
+    <NumberFormat
+      {...other}
+      getInputRef={inputRef}
+      onValueChange={values => {
+        onChange({
+          target: {
+            value: values.value
+          }
+        });
+      }}
+      thousandSeparator
+    />
+  );
+}
+
+export default applyNumberFormat;
