+++ frontend/src/pages/Overview/ProjectDialogContent.js
@@ -4,7 +4,7 @@ import Divider from "@material-ui/core/Divider";
 
 import strings from "../../localizeStrings";
 import ImageSelector from "../Common/ImageSelector";
-import Budget from "../Common/Budget2";
+import Budget from "../Common/Budget";
 import Identifier from "../Common/Identifier";
 import { toAmountString } from "../../helper";
