+++ frontend/src/pages/SubProjects/ProjectAssigneeContainer.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
+
+import { toJS } from "../../helper";
 import AssigneeSelection from "../Common/AssigneeSelection";
 import { assignProject } from "./actions";
-import withInitialLoading from "../Loading/withInitialLoading";
-import { toJS } from "../../helper";
 
 class ProjectAssigneeContainer extends Component {
   assignProject = identity => {
